api/serializers.py
from dataclasses import field
from rest_framework import serializers
from api.models import Author, Book

class AuthorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Author
        fields = ["id", "first_name", "last_name"]

class BookSerializer(serializers.ModelSerializer):
    author = AuthorSerializer(read_only=True)
    author_id = serializers.PrimaryKeyRelatedField(queryset=Author.objects.all(), write_only=True, source='author')
    class Meta:
        model = Book
        fields = ["id", "name", "isbn", "author", "author_id"]